WIP: T05 -- Reorganize backend telemetry test files
